Initial commit
This is the initial version—working as of June 4th—with landing page
nearly finished. Roulette is incorporated and working, header is almost
perfect. Everything else should be automatic. Might need to change some
html around to stop it prepending with site.base_url and thus not
displaying pictures and adding css/js.